refactor(render): use non "-desc" module for imports when appropriate
